+++ src/AddItem.jsx
@@ -1,0 +1,5 @@
+function AddItem() {
+  return <h2>임시 AddItem 페이지</h2>;
+}
+
+export default AddItem;
