+++ src/App.jsx
@@ -1,3 +1,6 @@
+import React, { Component } from "react";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+
 import { useState, useEffect } from "react";
 import Header from "/src/components/header/Header";
 import BestProductsList from "./components/BestProductsList";
@@ -8,69 +11,21 @@ import "./App.css";
 import AllProductsList from "./components/AllProductsList";
 import "/src/styles/Button.css";
 
+import Items from "./Items";
+import NotFound from "./NotFound";
+import AddItem from "./AddItem";
+
 function App() {
-  const [allProducts, setAllProducts] = useState([]);
-  const [bestProducts, setBestProducts] = useState([]);
-  const [order, setOrder] = useState("recent");
-  const [page, setPage] = useState(1);
-  const [pageSize, setPageSize] = useState(10);
-
-  const [windowSize, setWindowSize] = useState(window.innerWidth);
-
-  const sortedProducts = allProducts.sort((a, b) => b[order] - a[order]);
-
-  const handleBestProductsLoad = async (options) => {
-    const { list } = await getProducts(options);
-    setBestProducts(list);
-  };
-
-  const handleAllProductsLoad = async (options) => {
-    const { list } = await getProducts(options);
-    setAllProducts(list);
-  };
-
-  window.onresize = () => {
-    setWindowSize(window.innerWidth);
-
-    if (windowSize >= 1200) setPageSize(10);
-    else if (1200 > windowSize && windowSize >= 768) setPageSize(6);
-    else setPageSize(4);
-  };
-
-  useEffect(() => {
-    handleBestProductsLoad({
-      orderBy: "favorite",
-      pageSize: 4,
-      page: 1,
-    });
-  }, []);
-
-  useEffect(() => {
-    handleAllProductsLoad({
-      orderBy: `${order}`,
-      pageSize: `${pageSize}`,
-      page: `${page}`,
-    });
-  }, [order, page, pageSize]);
-
   return (
     <>
-      <Header id="header" />
-      <main id="main">
-        <BestProductsList
-          className="best-products-list"
-          products={bestProducts}
-        />
-        <AllProductsList
-          handleProductsLoad={handleAllProductsLoad}
-          page={page}
-          setPage={setPage}
-          order={order}
-          setOrder={setOrder}
-          className="all-products-list"
-          products={sortedProducts}
-        />
-      </main>
+      <BrowserRouter>
+        <Header className="header" />
+        <Routes>
+          <Route path="/items" element={<Items />}></Route>
+          <Route path="*" element={<NotFound />}></Route>
+          <Route path="/additem" element={<AddItem />}></Route>
+        </Routes>
+      </BrowserRouter>
     </>
   );
 }
@@ -78,5 +33,4 @@ function App() {
 export default App;
 
 // TODO: 반응형 모바일 로고 변경
-// TODO: 페이지네이션 구현
 // TODO: 중고마켓 눌렀을때 items로 이동
