+++ src/NotFound.jsx
@@ -1,0 +1,5 @@
+function NotFound() {
+  return <h2>Not Found</h2>;
+}
+
+export default NotFound;
