+++ src/components/common/Button.jsx
@@ -1,5 +1,9 @@
-function Button({ className, children }) {
-  return <button className={className}>{children}</button>;
+function Button({ className, children, onClick }) {
+  return (
+    <button onClick={onClick} className={className}>
+      {children}
+    </button>
+  );
 }
 
 export default Button;
