+++ src/components/header/TabMenu.jsx
@@ -14,6 +14,7 @@ function TabMenu({ className }) {
       <TabMenuList className="header__tab-list active" src="/items">
         중고마켓
       </TabMenuList>
+      {/* TODO: NavList로 바꿔서 깨진 css 정리하개 */}
     </div>
   );
 }
