+++ src/components/header/TabMenuList.jsx
@@ -1,10 +1,20 @@
+import React from "react";
+import { NavLink } from "react-router-dom";
 import "/src/styles/header/TabMenuList.css";
+
+const activeStyle = {
+  color: "#3692FF",
+};
 
 function TabMenuList({ className, children, src }) {
   return (
-    <li className={className}>
-      <a href={src}>{children}</a>
-    </li>
+    <NavLink
+      className={`link ${className}`}
+      style={({ isActive }) => (isActive ? activeStyle : {})}
+      to={src}
+    >
+      {children}
+    </NavLink>
   );
 }
